app/requirements.txt
blinker
simplejson
watchdog
flask
requests
oauthlib
pyOpenSSL
flask-login
google-cloud-datastore
